build(deps): update dependencies and fix adapter issues
- Update AGP to 8.11.0-alpha03
- Update core from1.16.0-beta01 to 1.16.0-rc01
- Update customview from 1.2.0-alpha02 to 1.2.0-beta01
- Update dynamicanimation from 1.1.0-beta01 to 1.1.0-rc01
- Update gridlayout from 1.1.0-beta01 to 1.1.0-rc01
- Fix getItemViewType in adapters to return Int instead of Long
- Improve error handling in PopupRecyclerAdapter and PopupRecyclerListAdapter

+++ app/src/main/java/github/znzsofficial/adapter/LuaCustRecyclerAdapter.kt
@@ -22,7 +22,7 @@ open class LuaCustRecyclerAdapter @JvmOverloads constructor(
 
     override fun getItemViewType(i: Int): Int {
         return try {
-            adapterCreator.getItemViewType(i).toInt()
+            adapterCreator.getItemViewType(i)
         } catch (e: Exception) {
             mContext?.sendError("RecyclerAdapter: getItemViewType", e)
             -1
@@ -57,7 +57,7 @@ open class LuaCustRecyclerAdapter @JvmOverloads constructor(
 
     interface Creator {
         fun getItemCount(): Int
-        fun getItemViewType(i: Int): Long
+        fun getItemViewType(i: Int): Int
         fun onBindViewHolder(viewHolder: LuaCustRecyclerHolder, i: Int)
         fun onCreateViewHolder(viewGroup: ViewGroup?, i: Int): LuaCustRecyclerHolder
         fun onViewRecycled(viewHolder: LuaCustRecyclerHolder)
